refactored test booking user methods to traits
